Extract GraphQL schema out of route
