+++ server/routes/index.ts
@@ -1,19 +1,5 @@
-import { queryType, stringArg, makeSchema } from "nexus";
 import { ApolloServer } from "apollo-server-micro";
-
-const Query = queryType({
-  definition(t) {
-    t.string("hello", {
-      args: { name: stringArg({ nullable: true }) },
-      resolve: (_, { name }) => `Hello ${name || "World"}!`
-    });
-  }
-});
-
-const schema = makeSchema({
-  types: [Query],
-  outputs: false
-});
+import schema from "../graphql/schema";
 
 const server = new ApolloServer({
   schema,
